# Increase parsing performance