This PR will increase parsing performance slightly.

−               | before   | after   | result
--               | --       | --      | --
parse            | 66.675   | 70.205  | 1.053x
render           | 225.502  | 225.397 | -
parse & render   | 48.816   | 50.937  | 1.043x


### Environment
- MacBook Air (M1, 2020)
- macOS 12.0 beta 7
- Apple M1
- Ruby 3.0.2

### Before
```
$ rake benchmark:run
/Users/watson/.rbenv/versions/3.0.2/bin/ruby ./performance/benchmark.rb lax

Running benchmark for 10 seconds (with 5 seconds warmup).

Warming up --------------------------------------
              parse:     6.000  i/100ms
             render:    22.000  i/100ms
     parse & render:     4.000  i/100ms
Calculating -------------------------------------
              parse:     66.675  (± 0.0%) i/s -    672.000  in  10.078787s
             render:    225.502  (± 0.9%) i/s -      2.266k in  10.049823s
     parse & render:     48.816  (± 0.0%) i/s -    492.000  in  10.078850s
```

### After
```
$ rake benchmark:run
/Users/watson/.rbenv/versions/3.0.2/bin/ruby ./performance/benchmark.rb lax

Running benchmark for 10 seconds (with 5 seconds warmup).

Warming up --------------------------------------
              parse:     6.000  i/100ms
             render:    21.000  i/100ms
     parse & render:     5.000  i/100ms
Calculating -------------------------------------
              parse:     70.205  (± 0.0%) i/s -    708.000  in  10.084957s
             render:    225.397  (± 0.9%) i/s -      2.268k in  10.063268s
     parse & render:     50.937  (± 0.0%) i/s -    510.000  in  10.012423s
```